refactor: Move group status wait to wait.rs
