+++ tests/helper/wait.rs
@@ -7,6 +7,7 @@
 use anyhow::{bail, Result};
 
 use pueue_lib::settings::Shared;
+use pueue_lib::state::GroupStatus;
 use pueue_lib::task::{Task, TaskStatus};
 
 use super::{get_state, sleep_ms};
@@ -129,4 +130,31 @@ pub async fn wait_for_group_absence(shared: &Shared, group: &str) -> Result<()> 
     }
 
     bail!("Group {group} hasn't been removed after about 1 second.")
+}
+
+/// Waits for a status on a specific group.
+pub async fn wait_for_group_status(
+    shared: &Shared,
+    group: &str,
+    expected_status: GroupStatus,
+) -> Result<()> {
+    let state = get_state(shared).await?;
+
+    // Give the daemon about 1 sec to shutdown.
+    let tries = 20;
+    let mut current_try = 0;
+
+    while current_try < tries {
+        // Process is still alive, wait a little longer
+        if let Some(status) = state.groups.get(group) {
+            if matches!(status, _expected_status) {
+                return Ok(());
+            }
+        }
+
+        sleep_ms(50);
+        current_try += 1;
+    }
+
+    bail!("Group {group} didn't change to state {expected_status:?} after about 1 sec.",);
 }
